pyproject.toml
[project]
name = "devops-copilot"
version = "0.1.0"
description = "AI DevOps Copilot — LLM-powered incident detection, diagnosis, and remediation"
authors = [
    {name = "Antigravity",email = "antigravity@google.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pydantic>=2.0",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "chromadb>=0.4",
    "prometheus-client>=0.17",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "httpx>=0.24",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "simpleeval>=1.0.3,<2.0.0",
    "google-genai (>=1.64.0,<2.0.0)",
    "groq (>=1.0.0,<2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]

[tool.poetry]
packages = [{include = "devops_copilot"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
